+++ src/main/java/com/classroom/equipment/config/Constants.java
@@ -1,0 +1,5 @@
+package com.classroom.equipment.config;
+
+public class Constants {
+    public static final String EMAIL_NOTIFICATION_TEMPLATE = "email-notification";
+}