+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGE LOG
 All notable changes to this project will be documented in this file.
 
+## feature-dev (2024-10-20)
+- Thêm module gửi email
+
 ## feature-dev (2024-10-10)
 - Khởi tạo project